+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - Add `device` and `device_move` methods on backend with universal device string representation
 
-- Add `to_qir` and `to_circuit` method for `DMCircuit` class
+- Add `to_qir`, `from_qir`, `to_circuit`, `inverse` method for `DMCircuit` class
 
 ### Changed